internal: Move internal-only helper methods to internal package
